잘못된 변수명 및 함수명, 변수 타입 수정

+++ app/src/main/java/com/lifekau/android/lifekau/TotalCurrGrade.java
@@ -1,7 +1,7 @@
 public class TotalCurrGrade {
     int registeredCredits;
     int acquiredCredits;
-    int totalGrades;
+    double totalGrades;
     double GPA;
     int semesterRanking;
     String remarks;
